requirements.txt
pydantic>=2.0.0
pdfplumber>=0.10.0
docling>=1.0.0
rapidocr-onnxruntime>=1.3.0
pyyaml>=6.0.0
httpx>=0.24.0
pytest>=7.0.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0